Removed code related to the Post-Provisioning Inventory swap.
While the playbook is able to delete and host and then add the host to
a different inventory, it will only be able to delete a host when the
inventory the host is in does not have any current jobs running against
it.  Therefore the swap will fail when performing ZTP on more than one
device at a time.